requirements.txt
google-genai
python-dotenv
pandas
ipykernel
transformers
xgboost
nltk
sklearn